# add membership test methods to tuples