@bicycle1885 2018-02-18
Thank you. I've set an upper limit to avoid the blow-up.