@chethega 2018-02-18
```
3 in (missing,4,5,3)
ERROR: TypeError: non-boolean (Missing) used in boolean context
```