@nalimilan 2018-02-18
Good catch @chakravala. Should probably use `invoke` to call the generic `in` method here.