@chethega 2018-02-18
The current state of the PR already calls, essentially, the old version. Hence, `-1 in ((1:100)...,missing) ==missing` (but the invoked method appears to be faster, for whatever reason. Maybe I just failed at testing). So one would only need to reproduce the correct missing behavior for small tuples.

Still has quadratic runtime for large inhomogeneous tuples, but I am seriously unsure how to avoid this.