@bicycle1885 2018-03-18
master:
```
julia> @benchmark count(foo, x)
┌ Warning: `indmin` is deprecated, use `argmin` instead.
│   caller = minimum at trials.jl:112 [inlined]
└ @ Core trials.jl:112
┌ Warning: `indmax` is deprecated, use `argmax` instead.
│   caller = maximum at trials.jl:117 [inlined]
└ @ Core trials.jl:117
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     3.883 μs (0.00% GC)
  median time:      3.948 μs (0.00% GC)
  mean time:        4.366 μs (0.00% GC)
  maximum time:     74.570 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     8

```

this pull request:
```
julia> srand(1234); x = rand(1:100, 1000);

julia> @benchmark count(foo, x)
┌ Warning: `indmin` is deprecated, use `argmin` instead.
│   caller = minimum at trials.jl:112 [inlined]
└ @ Core trials.jl:112
┌ Warning: `indmax` is deprecated, use `argmax` instead.
│   caller = maximum at trials.jl:117 [inlined]
└ @ Core trials.jl:117
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     925.417 ns (0.00% GC)
  median time:      1.033 μs (0.00% GC)
  mean time:        1.149 μs (0.00% GC)
  maximum time:     10.934 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     12

julia> @benchmark count(bar, x)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     940.739 ns (0.00% GC)
  median time:      1.026 μs (0.00% GC)
  mean time:        1.107 μs (0.00% GC)
  maximum time:     8.635 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     23

```

So, there is no detectable performance difference between `x in tuple` and `x == tuple[1] || x == tuple[2] || ...` for short tuples.