@StefanKarpinski 2018-03-20
Nice. `x in tuple` is a much nicer idiom.